gis/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pyyaml==6.0.1
aiofiles==23.2.1
pymysql==1.1.1
